Rename to Datadog::DistributedTracing::Headers
