--- lib/ddtrace/propagation/distributed_headers/.#base.rb
@@ -1,1 +1,0 @@
-brett.langdon@COMP10665.local.53629
