+++ spec/ddtrace/distributed_tracing/headers/datadog_spec.rb
@@ -3,9 +3,9 @@ require 'spec_helper'
 require 'ddtrace'
 require 'ddtrace/context'
 require 'ddtrace/ext/distributed'
-require 'ddtrace/propagation/distributed_headers/datadog'
+require 'ddtrace/distributed_tracing/headers/datadog'
 
-RSpec.describe Datadog::DistributedHeaders::Datadog do
+RSpec.describe Datadog::DistributedTracing::Headers::Datadog do
   let(:context) { Datadog.Context.new }
 
   # Helper to format env header keys
